refactor(supernova): name change
