+++ apps/supernova/src/AppContent.jsx
@@ -38,7 +38,6 @@ const AppContent = () => {
 
   useEffect(() => {
     if (silencesData) {
-      console.log(silencesData)
       setSilences({
         items: silencesData?.silences,
       })
